Cleaned codebase, created components folder
