+++ client/src/app/page.tsx
@@ -6,8 +6,10 @@
 */
 import Link from "next/link";
 import Image from "next/image";
-import { useEffect, useState, FormEvent, useRef } from "react";
+import { useEffect, useState } from "react";
 import { getSupabaseClient } from "@/lib/supabaseClient";
+import ProfileDropdown from "@/components/layout/ProfileDropdown";
+import ContactForm from "@/components/ui/ContactForm";
 
 /**
  * Landing page route component.
@@ -18,21 +20,6 @@ import { getSupabaseClient } from "@/lib/supabaseClient";
  */
 export default function Home() {
   const [isAuthed, setIsAuthed] = useState<boolean | null>(null);
-  
-  // Contact form submit handler (client-side placeholder)
-  function handleContactSubmit(event: FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    const formData = new FormData(event.currentTarget);
-    const name = String(formData.get("name") || "");
-    const email = String(formData.get("email") || "");
-    const message = String(formData.get("message") || "");
-    if (!name || !email || !message) {
-      alert("Please complete all fields.");
-      return;
-    }
-    alert("Thanks! Your message has been received.");
-    (event.currentTarget as HTMLFormElement).reset();
-  }
 
   // Auth state bootstrap and live sync:
   // - Check session (fast path), fallback to getUser
@@ -98,7 +85,7 @@ export default function Home() {
           {/* Right: Auth actions */}
           <div className="justify-self-end">
             {isAuthed ? (
-              <LandingPageProfileDropdown />
+              <ProfileDropdown variant="landing" />
             ) : (
               <Link
                 href="/login"
@@ -160,24 +147,7 @@ export default function Home() {
       {/* Contact Us */}
       <section className="border-t border-b/50 bg-white/30">
         <div className="mx-auto w-full max-w-screen-2xl px-6 lg:px-8 py-16">
-          <h2 className="text-2xl font-semibold text-center">Contact Us</h2>
-          <form onSubmit={handleContactSubmit} className="mt-8 grid gap-4 max-w-xl mx-auto">
-            <div>
-              <label className="block text-sm font-medium">Name</label>
-              <input name="name" type="text" required className="mt-1 w-full rounded-md border px-3 py-2 outline-none focus:ring-2 focus:ring-black" />
-            </div>
-            <div>
-              <label className="block text-sm font-medium">Email</label>
-              <input name="email" type="email" required className="mt-1 w-full rounded-md border px-3 py-2 outline-none focus:ring-2 focus:ring-black" />
-            </div>
-            <div>
-              <label className="block text-sm font-medium">Message</label>
-              <textarea name="message" rows={5} required className="mt-1 w-full rounded-md border px-3 py-2 outline-none focus:ring-2 focus:ring-black" />
-            </div>
-            <div>
-              <button type="submit" className="inline-flex items-center rounded-md bg-black px-4 py-2 text-white hover:bg-gray-800 cursor-pointer">Send</button>
-            </div>
-          </form>
+          <ContactForm />
         </div>
       </section>
 
@@ -189,129 +159,4 @@ export default function Home() {
   );
 }
 
-/**
- * Generate initials from a user's name
- */
-function getInitials(name: string): string {
-  if (!name) return 'U';
-  
-  const words = name.trim().split(/\s+/);
-  if (words.length === 1) {
-    return words[0].charAt(0).toUpperCase();
-  } else {
-    return words.slice(0, 2).map(word => word.charAt(0).toUpperCase()).join('');
-  }
-}
 
-/**
- * Profile dropdown component for the landing page
- */
-function LandingPageProfileDropdown() {
-  const [isOpen, setIsOpen] = useState(false);
-  const [userName, setUserName] = useState<string | null>(null);
-  const [username, setUsername] = useState<string | null>(null);
-  const dropdownRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    // Get user info for display
-    const supabase = getSupabaseClient();
-    
-    // Get user metadata (name) and database info (username)
-    const fetchUserData = async () => {
-      try {
-        const { data } = await supabase.auth.getUser();
-        const user = data.user;
-        if (user) {
-          setUserName(user.user_metadata?.name ?? null);
-          
-          try {
-            // Get username from Users table
-            const { data: dbResult } = await supabase
-              .from('Users')
-              .select('username')
-              .eq('user_id', user.id)
-              .single();
-            
-            setUsername(dbResult?.username ?? user.email?.split('@')[0] ?? 'User');
-          } catch {
-            // Fallback to email prefix if database query fails
-            setUsername(user.email?.split('@')[0] ?? 'User');
-          }
-        }
-      } catch (error) {
-        console.error('Error fetching user data:', error);
-      }
-    };
-    
-    fetchUserData();
-
-    // Close dropdown when clicking outside
-    function handleClickOutside(event: MouseEvent) {
-      if (dropdownRef.current && !dropdownRef.current.contains(event.target as Node)) {
-        setIsOpen(false);
-      }
-    }
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, []);
-
-  async function handleSignOut() {
-    const supabase = getSupabaseClient();
-    await supabase.auth.signOut();
-    window.location.assign("/");
-  }
-
-  // Get initials for avatar
-  const initials = getInitials(userName || 'User');
-
-  return (
-    <div className="relative" ref={dropdownRef}>
-      {/* Avatar button */}
-      <button
-        onClick={() => setIsOpen(!isOpen)}
-        className="w-8 h-8 rounded-full bg-gray-600 text-white text-sm font-medium flex items-center justify-center hover:bg-gray-700 transition-colors"
-        aria-label="Profile menu"
-      >
-        {initials}
-      </button>
-
-      {/* Dropdown menu */}
-      {isOpen && (
-        <div className="absolute right-0 mt-2 w-48 bg-white border border-gray-200 rounded-md shadow-lg z-50">
-          <div className="py-1">
-            {/* User info */}
-            <div className="px-4 py-2 text-sm text-gray-700 border-b">
-              {username || 'User'}
-            </div>
-            
-            {/* Navigation links */}
-            <Link
-              href="/home"
-              className="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-              onClick={() => setIsOpen(false)}
-            >
-              Dashboard
-            </Link>
-            
-            <Link
-              href="/profile"
-              className="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-              onClick={() => setIsOpen(false)}
-            >
-              Profile
-            </Link>
-            
-            {/* Sign out */}
-            <button
-              onClick={handleSignOut}
-              className="block w-full text-left px-4 py-2 text-sm text-gray-700 hover:bg-gray-100"
-            >
-              Sign out
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
